# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Vigilância Participativa API
  description: API REST para sistema de vigilância participativa em saúde pública
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: https://staging-api.example.com/v1
    description: Staging server
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT no formato Bearer
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 100
        totalPages:
          type: integer
        totalItems:
          type: integer
    PaginationLinks:
      type: object
      properties:
        first:
          type: string
          format: uri
        last:
          type: string
          format: uri
        prev:
          type: string
          format: uri
          nullable: true
        next:
          type: string
          format: uri
          nullable: true
...
    Location:
      type: object
      required:
        - name
      properties:
        id:
          type: integer
          readOnly: true
        parentId:
          type: integer
          nullable: true
        name:
          type: string
        latitude:
          type: number
          format: double
          nullable: true
        longitude:
          type: number
          format: double
          nullable: true
        polygons:
          type: object
          nullable: true
        active:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
    LocationCreate:
      type: object
      required:
        - name
      properties:
        parentId:
          type: integer
          nullable: true
        name:
          type: string
        latitude:
          type: number
          format: double
          nullable: true
        longitude:
          type: number
          format: double
          nullable: true
        polygons:
          type: object
          nullable: true
        active:
          type: boolean
          default: true
    LocationUpdate:
      type: object
      properties:
        parentId:
          type: integer
          nullable: true
        name:
          type: string
        latitude:
          type: number
          format: double
          nullable: true
        longitude:
          type: number
          format: double
          nullable: true
        polygons:
          type: object
          nullable: true
        active:
          type: boolean
...
    Form:
      type: object
      required:
        - title
        - type
      properties:
        id:
          type: integer
          readOnly: true
        contextId:
          type: integer
          nullable: true
        title:
          type: string
        reference:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [signal, quiz]
        active:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
    FormCreate:
      type: object
      required:
        - title
        - type
      properties:
        contextId:
          type: integer
          nullable: true
        title:
          type: string
        reference:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [signal, quiz]
        active:
          type: boolean
          default: true
    FormUpdate:
      type: object
      properties:
        contextId:
          type: integer
          nullable: true
        title:
          type: string
        reference:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [signal, quiz]
        active:
          type: boolean
    FormVersion:
      type: object
      required:
        - formId
        - versionNumber
        - accessType
        - definition
      properties:
        id:
          type: integer
          readOnly: true
        formId:
          type: integer
        versionNumber:
          type: integer
        accessType:
          type: string
          enum: [PUBLIC, PRIVATE]
        definition:
          type: object
        active:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
    FormVersionCreate:
      type: object
      required:
        - formId
        - versionNumber
        - accessType
        - definition
      properties:
        formId:
          type: integer
        versionNumber:
          type: integer
        accessType:
          type: string
          enum: [PUBLIC, PRIVATE]
        definition:
          type: object
        active:
          type: boolean
          default: true
    FormVersionUpdate:
      type: object
      properties:
        formId:
          type: integer
        versionNumber:
          type: integer
        accessType:
          type: string
          enum: [PUBLIC, PRIVATE]
        definition:
          type: object
        active:
          type: boolean
...
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: ID do recurso
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Número da página
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Tamanho da página
  responses:
    BadRequestError:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnauthorizedError:
      description: Não autorizado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFoundError:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /locations:
    get:
      tags:
        - Locations
      summary: Listar localizações
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: active
          in: query
          schema:
            type: boolean
        - name: parentId
          in: query
          schema:
            type: integer
            nullable: true
      responses:
        '200':
          description: Lista de localizações
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Location'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
                  links:
                    $ref: '#/components/schemas/PaginationLinks'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Locations
      summary: Criar localização
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationCreate'
      responses:
        '201':
          description: Localização criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /locations/{id}:
    get:
      tags:
        - Locations
      summary: Obter localização
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Detalhes da localização
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    put:
      tags:
        - Locations
      summary: Atualizar localização
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationUpdate'
      responses:
        '200':
          description: Localização atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      tags:
        - Locations
      summary: Deletar localização
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Localização deletada
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /forms:
    get:
      tags:
        - Forms
      summary: Listar formulários
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: contextId
          in: query
          schema:
            type: integer
            nullable: true
        - name: type
          in: query
          schema:
            type: string
            enum: [signal, quiz]
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Lista de formulários
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Form'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
                  links:
                    $ref: '#/components/schemas/PaginationLinks'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Forms
      summary: Criar formulário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FormCreate'
      responses:
        '201':
          description: Formulário criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Form'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /forms/{id}:
    get:
      tags:
        - Forms
      summary: Obter formulário
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Detalhes do formulário
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Form'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    put:
      tags:
        - Forms
      summary: Atualizar formulário
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FormUpdate'
      responses:
        '200':
          description: Formulário atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Form'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      tags:
        - Forms
      summary: Deletar formulário
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Formulário deletado
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /forms/{formId}/versions:
    get:
      tags:
        - Form Versions
      summary: Listar versões do formulário
      parameters:
        - name: formId
          in: path
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Lista de versões
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/FormVersion'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
                  links:
                    $ref: '#/components/schemas/PaginationLinks'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Form Versions
      summary: Criar versão do formulário
      parameters:
        - name: formId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FormVersionCreate'
      responses:
        '201':
          description: Versão criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormVersion'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /forms/{formId}/versions/{id}:
    get:
      tags:
        - Form Versions
      summary: Obter versão do formulário
      parameters:
        - name: formId
          in: path
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Detalhes da versão
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormVersion'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    put:
      tags:
        - Form Versions
      summary: Atualizar versão do formulário
      parameters:
        - name: formId
          in: path
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FormVersionUpdate'
      responses:
        '200':
          description: Versão atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormVersion'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      tags:
        - Form Versions
      summary: Deletar versão do formulário
      parameters:
        - name: formId
          in: path
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Versão deletada
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...